# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Laravel E-Commerce API'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
paths:
...
  /api/v1/admin/account-settings/change-password:
    post:
      summary: 'Change Password'
      operationId: changePassword
      description: "Update the password of the currently authenticated admin user.\nIt will be updated only when the current password is matched."
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
      responses:
        201:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: success
                  message: 'Password updated successfully.'
                  data: []
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: 'Password updated successfully.'
                  data:
                    type: array
                    example: []
      tags:
        - 'Administrator Endpoints'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                current_password:
                  type: string
                  description: "The admin user's current password."
                  example: Pa$$w0rd
                new_password:
                  type: string
                  description: 'The new password.'
                  example: Secret
                new_password_confirmation:
                  type: string
                  description: 'Confirm the new password, should match new_password.'
                  example: Secret
              required:
                - current_password
                - new_password
                - new_password_confirmation
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: ''
security:
  -
    default: []